+++ src/barchartstacked/mapping.js
@@ -13,19 +13,15 @@ export const mapData = function (data, mapping, dataTypes, dimensions) {
     dimensions
   )
 
-  // add the non-compulsory dimensions.
-  'series' in mapping ? null : (mapping.series = { value: undefined })
-
   let results = []
   const result = d3.rollups(
     data,
     (v) => {
-      
+      // @TODO use the spread operator to creat groups on mapping values
       // for every dimension in the bars field, create an item
       mapping.bars.value.forEach((barName, i) => {
-        
         //getting values for aggregation
-        const valuesForSize = v.map(x => x[barName])
+        const valuesForSize = v.map((x) => x[barName])
         //getting i-th aggregator
         const aggregator = barsAggregators[i]
 
@@ -34,10 +30,7 @@ export const mapData = function (data, mapping, dataTypes, dimensions) {
           series: v[0][mapping.series.value], // get the first one since it's grouped
           stacks: v[0][mapping.stacks.value], // get the first one since it's grouped
           bars: barName,
-          // #TODO: remove me if dynamic aggregation works
-          // size: v.reduce((result, elm) => result + elm[barName], 0),
           size: aggregator(valuesForSize),
-          color: barName,
         }
         results.push(item)
       })