+++ src/barchartstacked/visualOptions.js
@@ -70,6 +70,19 @@ export const visualOptions = {
     default: 'None',
   },
 
+  SortXAxisBy: {
+    type: 'text',
+    label: 'Sort X axis by',
+    group: 'chart',
+    options: [
+      'Total value (descending)',
+      'Total value (ascending)',
+      'Name',
+      'Original',
+    ],
+    default: 'Name',
+  },
+
   useSameScale: {
     type: 'boolean',
     label: 'Use same scale',
@@ -104,10 +117,17 @@ export const visualOptions = {
     group: 'series',
   },
 
+  showGrid: {
+    type: 'boolean',
+    label: 'Show series grid',
+    default: true,
+    group: 'series',
+  },
+
   colorScale: {
     type: 'colorScale',
     label: 'Color scale',
-    dimension: 'color',
+    dimension: 'bars',
     default: {
       scaleType: 'ordinal',
       interpolator: 'interpolateSpectral',